+++ src/Trash/Radarr/CustomFormat/Models/UpdatedFormatScore.cs
@@ -1,0 +1,13 @@
+namespace Trash.Radarr.CustomFormat.Models
+{
+    public enum FormatScoreUpdateReason
+    {
+        Updated,
+        Reset
+    }
+
+    public record UpdatedFormatScore(
+        string CustomFormatName,
+        int Score,
+        FormatScoreUpdateReason Reason);
+}