+++ src/Trash/Radarr/CustomFormat/Processors/GuideSteps/IQualityProfileStep.cs
@@ -5,7 +5,7 @@ namespace Trash.Radarr.CustomFormat.Processors.GuideSteps
 {
     public interface IQualityProfileStep
     {
-        Dictionary<string, List<QualityProfileCustomFormatScoreEntry>> ProfileScores { get; }
+        Dictionary<string, QualityProfileCustomFormatScoreMapping> ProfileScores { get; }
         List<(string name, string trashId, string profileName)> CustomFormatsWithoutScore { get; }
         void Process(IEnumerable<ProcessedConfigData> configData);
     }