+++ src/Trash/Radarr/CustomFormat/Processors/GuideSteps/QualityProfileStep.cs
@@ -1,12 +1,11 @@
 using System.Collections.Generic;
-using Trash.Extensions;
 using Trash.Radarr.CustomFormat.Models;
 
 namespace Trash.Radarr.CustomFormat.Processors.GuideSteps
 {
     public class QualityProfileStep : IQualityProfileStep
     {
-        public Dictionary<string, List<QualityProfileCustomFormatScoreEntry>> ProfileScores { get; } = new();
+        public Dictionary<string, QualityProfileCustomFormatScoreMapping> ProfileScores { get; } = new();
         public List<(string name, string trashId, string profileName)> CustomFormatsWithoutScore { get; } = new();
 
         public void Process(IEnumerable<ProcessedConfigData> configData)
@@ -18,7 +17,7 @@ namespace Trash.Radarr.CustomFormat.Processors.GuideSteps
                 // Check if there is a score we can use. Priority is:
                 //      1. Score from the YAML config is used. If user did not provide,
                 //      2. Score from the guide is used. If the guide did not have one,
-                //      3. Warn the user and skip it.
+                //      3. Warn the user and
                 var scoreToUse = profile.Score;
                 if (scoreToUse == null)
                 {
@@ -32,11 +31,18 @@ namespace Trash.Radarr.CustomFormat.Processors.GuideSteps
                     }
                 }
 
-                if (scoreToUse != null)
+                if (scoreToUse == null)
                 {
-                    ProfileScores.GetOrCreate(profile.Name)
-                        .Add(new QualityProfileCustomFormatScoreEntry(cf, scoreToUse.Value));
+                    continue;
                 }
+
+                if (!ProfileScores.TryGetValue(profile.Name, out var mapping))
+                {
+                    mapping = new QualityProfileCustomFormatScoreMapping(profile.ResetUnmatchedScores);
+                    ProfileScores[profile.Name] = mapping;
+                }
+
+                mapping.Mapping.Add(new FormatMappingEntry(cf, scoreToUse.Value));
             }
         }
     }